modularize index
