+++ redisvl/index/__init__.py
@@ -1,0 +1,3 @@
+from redisvl.index.index import SearchIndex
+
+__all__ = ["SearchIndex"]
